requirements.txt
Django==4.0.3
django-allauth==0.50.0
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
Pillow==9.0.1
getsms
requests

